feat: add seeders and factories for all entities with sample data

+++ database/migrations/2025_03_27_192255_create_apprentices_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('email');
-            $table->integer('cell_number');
+            $table->bigInteger('cell_number');
             
             $table->unsignedBigInteger('course_id');
             $table->unsignedBigInteger('computer_id');
